整理相关方法
