--- elastic-fed/modules/havenask-engine/src/main/java/org/havenask/engine/index/engine/HavenaskSearchQueryPhase.java
@@ -1,2 +1,0 @@
-package org.havenask.engine.index.engine;public class HavenaskSearchQueryPhase {
-}
